Codechange: [Network] Move connection string parsing away from C-strings

+++ src/network/network_internal.h
@@ -119,6 +119,6 @@ StringID GetNetworkErrorMsg(NetworkErrorCode err);
 bool NetworkFindName(char *new_name, const char *last);
 const char *GenerateCompanyPasswordHash(const char *password, const char *password_server_id, uint32 password_game_seed);
 
-NetworkAddress ParseConnectionString(const std::string &connection_string, int default_port);
+NetworkAddress ParseConnectionString(const std::string &connection_string, uint16 default_port);
 
 #endif /* NETWORK_INTERNAL_H */
